Revert "Add SCIM2 feature."
