--- components/org.wso2.carbon.identity.provisioning.connector.scim/src/main/java/org/wso2/carbon/identity/provisioning/connector/scim/AbstractSCIMOutboundProvisioningConnector.java
@@ -1,69 +1,0 @@
-/*
- * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.carbon.identity.provisioning.connector.scim;
-
-import org.wso2.carbon.identity.application.common.model.Property;
-import org.wso2.carbon.identity.provisioning.AbstractOutboundProvisioningConnector;
-import org.wso2.carbon.identity.provisioning.IdentityProvisioningConstants;
-import org.wso2.carbon.identity.provisioning.IdentityProvisioningException;
-import org.wso2.charon.core.config.SCIMConfigConstants;
-
-public abstract class AbstractSCIMOutboundProvisioningConnector extends AbstractOutboundProvisioningConnector {
-
-    protected String userStoreDomainName;
-
-    @Override
-    public void init(Property[] provisioningProperties) throws IdentityProvisioningException {
-        if (provisioningProperties != null && provisioningProperties.length > 0) {
-            for (Property property : provisioningProperties) {
-                if (SCIMProvisioningConnectorConstants.SCIM_USER_EP.equals(property.getName())) {
-                    populateSCIMProvider(property, SCIMConfigConstants.ELEMENT_NAME_USER_ENDPOINT);
-                } else if (SCIMProvisioningConnectorConstants.SCIM_GROUP_EP.equals(property.getName())) {
-                    populateSCIMProvider(property, SCIMConfigConstants.ELEMENT_NAME_GROUP_ENDPOINT);
-                } else if (SCIMProvisioningConnectorConstants.SCIM_USERNAME.equals(property.getName())) {
-                    populateSCIMProvider(property, SCIMConfigConstants.ELEMENT_NAME_USERNAME);
-                } else if (SCIMProvisioningConnectorConstants.SCIM_PASSWORD.equals(property.getName())) {
-                    populateSCIMProvider(property, SCIMConfigConstants.ELEMENT_NAME_PASSWORD);
-                } else if (SCIMProvisioningConnectorConstants.SCIM_USERSTORE_DOMAIN.equals(property.getName())) {
-                    userStoreDomainName = property.getValue() != null ? property.getValue()
-                            : property.getDefaultValue();
-                } else if (SCIMProvisioningConnectorConstants.SCIM_ENABLE_PASSWORD_PROVISIONING.equals(property.
-                        getName())) {
-                    populateSCIMProvider(property, SCIMProvisioningConnectorConstants.SCIM_ENABLE_PASSWORD_PROVISIONING);
-                } else if (SCIMProvisioningConnectorConstants.SCIM_DEFAULT_PASSWORD.equals(property.getName())) {
-                    populateSCIMProvider(property, SCIMProvisioningConnectorConstants.SCIM_DEFAULT_PASSWORD);
-                }
-
-                if (IdentityProvisioningConstants.JIT_PROVISIONING_ENABLED.equals(property
-                        .getName()) && "1".equals(property.getValue())) {
-                    jitProvisioningEnabled = true;
-                }
-            }
-        }
-    }
-
-    /**
-     * @param property
-     * @param scimPropertyName
-     * @throws IdentityProvisioningException
-     */
-    protected abstract void populateSCIMProvider(Property property, String scimPropertyName)
-            throws IdentityProvisioningException;
-
-}
