+++ components/org.wso2.carbon.identity.provisioning.connector.scim/src/main/java/org/wso2/carbon/identity/provisioning/connector/scim/SCIMProvisioningConnector.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.identity.provisioning.connector.scim.scim1;
+package org.wso2.carbon.identity.provisioning.connector.scim;
 
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.collections.MapUtils;
@@ -25,6 +25,7 @@ import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.wso2.carbon.identity.application.common.model.ClaimMapping;
 import org.wso2.carbon.identity.application.common.model.Property;
+import org.wso2.carbon.identity.provisioning.AbstractOutboundProvisioningConnector;
 import org.wso2.carbon.identity.provisioning.IdentityProvisioningConstants;
 import org.wso2.carbon.identity.provisioning.IdentityProvisioningException;
 import org.wso2.carbon.identity.provisioning.ProvisionedIdentifier;
@@ -32,8 +33,6 @@ import org.wso2.carbon.identity.provisioning.ProvisioningEntity;
 import org.wso2.carbon.identity.provisioning.ProvisioningEntityType;
 import org.wso2.carbon.identity.provisioning.ProvisioningOperation;
 import org.wso2.carbon.identity.provisioning.ProvisioningUtil;
-import org.wso2.carbon.identity.provisioning.connector.scim.AbstractSCIMOutboundProvisioningConnector;
-import org.wso2.carbon.identity.provisioning.connector.scim.SCIMProvisioningConnectorConstants;
 import org.wso2.carbon.identity.scim.common.impl.ProvisioningClient;
 import org.wso2.carbon.identity.scim.common.utils.AttributeMapper;
 import org.wso2.carbon.identity.scim.common.utils.SCIMCommonConstants;
@@ -51,11 +50,45 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
-public class SCIMProvisioningConnector extends AbstractSCIMOutboundProvisioningConnector {
+public class SCIMProvisioningConnector extends AbstractOutboundProvisioningConnector {
 
     private static final long serialVersionUID = -2800777564581005554L;
     private static Log log = LogFactory.getLog(SCIMProvisioningConnector.class);
-    private SCIMProvider scimProvider = new SCIMProvider();
+    private SCIMProvider scimProvider;
+    private String userStoreDomainName;
+
+    @Override
+    public void init(Property[] provisioningProperties) throws IdentityProvisioningException {
+        scimProvider = new SCIMProvider();
+
+        if (provisioningProperties != null && provisioningProperties.length > 0) {
+
+            for (Property property : provisioningProperties) {
+
+                if (SCIMProvisioningConnectorConstants.SCIM_USER_EP.equals(property.getName())) {
+                    populateSCIMProvider(property, SCIMConfigConstants.ELEMENT_NAME_USER_ENDPOINT);
+                } else if (SCIMProvisioningConnectorConstants.SCIM_GROUP_EP.equals(property.getName())) {
+                    populateSCIMProvider(property, SCIMConfigConstants.ELEMENT_NAME_GROUP_ENDPOINT);
+                } else if (SCIMProvisioningConnectorConstants.SCIM_USERNAME.equals(property.getName())) {
+                    populateSCIMProvider(property, SCIMConfigConstants.ELEMENT_NAME_USERNAME);
+                } else if (SCIMProvisioningConnectorConstants.SCIM_PASSWORD.equals(property.getName())) {
+                    populateSCIMProvider(property, SCIMConfigConstants.ELEMENT_NAME_PASSWORD);
+                } else if (SCIMProvisioningConnectorConstants.SCIM_USERSTORE_DOMAIN.equals(property.getName())) {
+                    userStoreDomainName = property.getValue() != null ? property.getValue()
+                            : property.getDefaultValue();
+                }else if (SCIMProvisioningConnectorConstants.SCIM_ENABLE_PASSWORD_PROVISIONING.equals(property.getName())){
+                    populateSCIMProvider(property, SCIMProvisioningConnectorConstants.SCIM_ENABLE_PASSWORD_PROVISIONING);
+                }else if (SCIMProvisioningConnectorConstants.SCIM_DEFAULT_PASSWORD.equals(property.getName())){
+                    populateSCIMProvider(property, SCIMProvisioningConnectorConstants.SCIM_DEFAULT_PASSWORD);
+                }
+
+                if (IdentityProvisioningConstants.JIT_PROVISIONING_ENABLED.equals(property
+                        .getName()) && "1".equals(property.getValue())) {
+                    jitProvisioningEnabled = true;
+                }
+            }
+        }
+    }
 
     @Override
     public ProvisionedIdentifier provision(ProvisioningEntity provisioningEntity)
@@ -413,6 +446,21 @@ public class SCIMProvisioningConnector extends AbstractSCIMOutboundProvisioningC
         return userStoreDomainName;
     }
 
+    /**
+     * @param property
+     * @param scimPropertyName
+     * @throws IdentityProvisioningException
+     */
+    private void populateSCIMProvider(Property property, String scimPropertyName)
+            throws IdentityProvisioningException {
+
+        if (property.getValue() != null && property.getValue().length() > 0) {
+            scimProvider.setProperty(scimPropertyName, property.getValue());
+        } else if (property.getDefaultValue() != null) {
+            scimProvider.setProperty(scimPropertyName, property.getDefaultValue());
+        }
+    }
+
     @Override
     public String getClaimDialectUri() throws IdentityProvisioningException {
         return SCIMProvisioningConnectorConstants.DEFAULT_SCIM_DIALECT;
@@ -427,17 +475,6 @@ public class SCIMProvisioningConnector extends AbstractSCIMOutboundProvisioningC
             user.setPassword(getPassword(userEntity.getAttributes()));
         } else if (StringUtils.isNotBlank(scimProvider.getProperty(SCIMProvisioningConnectorConstants.SCIM_DEFAULT_PASSWORD))) {
             user.setPassword(scimProvider.getProperty(SCIMProvisioningConnectorConstants.SCIM_DEFAULT_PASSWORD));
-        }
-    }
-
-    @Override
-    protected void populateSCIMProvider(Property property, String scimPropertyName)
-            throws IdentityProvisioningException {
-
-        if (property.getValue() != null && property.getValue().length() > 0) {
-            scimProvider.setProperty(scimPropertyName, property.getValue());
-        } else if (property.getDefaultValue() != null) {
-            scimProvider.setProperty(scimPropertyName, property.getDefaultValue());
         }
     }
