+++ components/org.wso2.carbon.identity.provisioning.connector.scim/src/main/java/org/wso2/carbon/identity/provisioning/connector/scim/SCIMProvisioningConnectorFactory.java
@@ -23,8 +23,6 @@ import org.apache.commons.logging.LogFactory;
 import org.wso2.carbon.identity.application.common.model.Property;
 import org.wso2.carbon.identity.provisioning.AbstractProvisioningConnectorFactory;
 import org.wso2.carbon.identity.provisioning.IdentityProvisioningException;
-import org.wso2.carbon.identity.provisioning.connector.scim.scim1.SCIMProvisioningConnector;
-import org.wso2.carbon.identity.provisioning.connector.scim.scim2.SCIM2ProvisioningConnector;
 
 /**
  * @author
@@ -32,30 +30,19 @@ import org.wso2.carbon.identity.provisioning.connector.scim.scim2.SCIM2Provision
 public class SCIMProvisioningConnectorFactory extends AbstractProvisioningConnectorFactory {
 
     public static final String SCIM = "scim";
-    private static final String SCIM_VERSION1 = "scim1";
-    private static final String SCIM_VERSION2 = "scim2";
     private static final Log log = LogFactory.getLog(SCIMProvisioningConnectorFactory.class);
 
     @Override
-    protected AbstractSCIMOutboundProvisioningConnector buildConnector(Property[] provisioningProperties)
+    /**
+     * @throws IdentityProvisioningException
+     */
+    protected SCIMProvisioningConnector buildConnector(Property[] provisioningProperties)
             throws IdentityProvisioningException {
-        String scimVersion = SCIM_VERSION1;
-        for (Property property : provisioningProperties) {
-            if (SCIMProvisioningConnectorConstants.SCIM_VERSION.equals(property.getName())) {
-                scimVersion = property.getValue();
-            }
-        }
-
-        AbstractSCIMOutboundProvisioningConnector scimProvisioningConnector;
-        if (scimVersion.equalsIgnoreCase(SCIM_VERSION1)) {
-            scimProvisioningConnector = new SCIMProvisioningConnector();
-        } else {
-            scimProvisioningConnector = new SCIM2ProvisioningConnector();
-        }
-
+        SCIMProvisioningConnector scimProvisioningConnector = new SCIMProvisioningConnector();
         scimProvisioningConnector.init(provisioningProperties);
+
         if (log.isDebugEnabled()) {
-            log.debug("Created new connector of type : " + scimVersion);
+            log.debug("Created new connector of type : " + SCIM);
         }
         return scimProvisioningConnector;
     }
@@ -64,4 +51,5 @@ public class SCIMProvisioningConnectorFactory extends AbstractProvisioningConnec
     public String getConnectorType() {
         return SCIM;
     }
+
 }
