--- components/org.wso2.carbon.identity.provisioning.connector.scim/src/main/java/org/wso2/carbon/identity/provisioning/connector/scim/scim2/SCIMCommonConstants.java
@@ -1,26 +1,0 @@
-/*
- * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.wso2.carbon.identity.provisioning.connector.scim.scim2;
-
-public class SCIMCommonConstants {
-
-    public static final String DEFAULT = "default";
-    public static final int USER = 1;
-    public static final int GROUP = 2;
-    public static final String SCIM_CORE_CLAIM_DIALECT = "urn:ietf:params:scim:schemas:core:2.0";
-    public static final String SCIM_USER_CLAIM_DIALECT = "urn:ietf:params:scim:schemas:core:2.0:User";
-}
