+++ pom.xml
@@ -85,16 +85,6 @@
                 <version>${charon.wso2.version.identity}</version>
             </dependency>
             <dependency>
-                <groupId>org.wso2.charon</groupId>
-                <artifactId>org.wso2.charon3.core</artifactId>
-                <version>${charon3.wso2.version.identity}</version>
-            </dependency>
-            <dependency>
-                <groupId>org.wso2.carbon.extension.identity.scim2</groupId>
-                <artifactId>org.wso2.carbon.extension.identity.scim2.client</artifactId>
-                <version>${identity.scim2.client.version}</version>
-            </dependency>
-            <dependency>
                 <groupId>commons-lang.wso2</groupId>
                 <artifactId>commons-lang</artifactId>
                 <version>${commons-lang.wso2.version}</version>
@@ -283,12 +273,7 @@
 
         <!-- Charon version -->
         <charon.wso2.version.identity>2.1.1</charon.wso2.version.identity>
-        <charon3.wso2.version.identity>3.0.0</charon3.wso2.version.identity>
         <charon.core.imp.pkg.version.range>[2.0.1,3.0.0)</charon.core.imp.pkg.version.range>
-        <charon3.core.imp.pkg.version.range>[3.0.0,4.0.0)</charon3.core.imp.pkg.version.range>
-
-        <!-- SCIM 2.0 client -->
-        <identity.scim2.client.version>[1.0.0,2.0.0)</identity.scim2.client.version>
 
         <!-- Commons -->
         <commons-lang.wso2.osgi.version.range>[2.6.0,3.0.0)</commons-lang.wso2.osgi.version.range>
